Begin instrumentation impl, remove log dep
This commit contains the beginning of the database instrumentation binary - this includes a DAO, some changes to the database models, and to the schema/models/db modules.

We also remove the dependency on 'log' in favor of an entirely 'slog' solution.  We set a global logging scope and provide that to the entirety of our binary.